Unnecessary long delay during receiving.
The reason why you needed to include the 15mS timer, is because the while loop executes very fast and the data from the Ikea module will be received slowly. Thus the while loop exits  too early if you do not had the 15mS delay included because there was no data (yet) inside the UART buffer. 

But your 15mS delay causes that for every received character, there will be 15mS delay. So 64 characters received will be 64 times 15mS equals a a long (blocking) code.

It's better to use this code, where I use a timeout for the entire frame and not for every character.