Fix up and document delay() hack in uart receive
This is still a hack but a little bit less of one with an actual
explanation instead of a scary comment.

The correct thing is to do is advance through a state machine that detects
the header and produces a symbol once it receives an entire message. But
all my hardware is currently deployed with this code so we'll ship
what's tested as opposed to what's theoretically correct.